Fix sorting of InstanceIterators, fixes #49
Before this commit, "empty" iterators we sorted at the end which resulted in the fast path for single iterators never being entered.
The resuls were still correct, but it didn't remove the depleted iterators from the array, which caused some overhead.

This also fixes how the exception cache is filled when multiple exception sources are present.