Split ByDayFilter, implements #52 (#84)
The original `ByDayFilter` was way to complex and hard to maintain. This commit breaks it into two much simpler classes. A part of the heavy lifting is now done in `RecurrenceRule` and should be subject to another refactoring.
A nice side effect is a small performance improvement when filtering `ByDay`.